+++ diffusion_policy/workspace/base_flow_matching_multi_gpu.py
@@ -49,13 +49,18 @@ def seed_worker(worker_id):
 class TrainingWorkspace(BaseWorkspace):
     include_keys = ['global_step', 'epoch']
 
-    def __init__(self, cfg: OmegaConf, output_dir=None):
+    def __init__(self, cfg: OmegaConf, output_dir=None, cfg_unresolved=None):
         super().__init__(cfg, output_dir=output_dir)
 
         # dump current config to yaml file:
         if dist.get_rank() == 0:
-            with open(self.output_dir + "/config_raw.yaml", "w") as f:
-                OmegaConf.save(cfg, f)
+            # dump current config to yaml file:
+            if not os.path.exists(self.output_dir + "/config_raw.yaml"):
+                with open(self.output_dir + "/config_raw.yaml", "w") as f:
+                    if cfg_unresolved is not None:
+                        OmegaConf.save(cfg_unresolved, f)
+                    else:
+                        OmegaConf.save(cfg, f)
 
         # configure model
         self.model = hydra.utils.instantiate(cfg.policy)
